+++ example/lib/pages/forces/forces_gem_bouncing.dart
@@ -2,7 +2,7 @@ import 'package:bonfire/bonfire.dart';
 import 'package:example/pages/mini_games/platform/platform_spritesheet.dart';
 
 class ForcesGemBouncing extends GameDecoration
-    with Movement, HandleForces, BlockMovementCollision, BouncingObject {
+    with Movement, HandleForces, BlockMovementCollision, ElasticCollision {
   ForcesGemBouncing({
     required Vector2 position,
   }) : super.withAnimation(
@@ -16,6 +16,7 @@ class ForcesGemBouncing extends GameDecoration
         value: Vector2(0, 100),
       ),
     );
+    setupElasticCollision(restitution: 3);
   }
 
   @override